Prevent empy Drip tags property from causing issues
There is no obvious way to debug Drip (although we think it's probably possible to set a debug attribute somewhere), however it is throwing an error sporadically on identify (related to an undefined array). This appears to be the only array that we are potentially defining, so defaulting the property may resolve issues.